pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ifc_structural_mechanics"
version = "0.1.0"
description = "Structural analysis functionality IFC using CalculiX"
readme = "README.md"
authors = [
    {name = "Bruno Postle", email = "bruno@postle.net"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "ifcopenshell>=0.7.0",
    "gmsh>=4.9.0",
    "meshio>=5.0.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "pandas>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=4.0.0",
    "mypy>=0.9.0",
]
docs = [
    "mkdocs>=1.3.0",
    "mkdocs-material>=8.0.0",
    "mkdocstrings>=0.18.0",
    "mkdocstrings-python>=0.7.0",
]

[project.urls]
"Homepage" = "https://github.com/example/ifc_structural_mechanics"
"Bug Tracker" = "https://github.com/example/ifc_structural_mechanics/issues"
"Documentation" = "https://example.github.io/ifc_structural_mechanics"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[project.scripts]
ifc-analysis = "ifc_structural_mechanics.cli:cli"

